requirements.txt
# OCR Medication Recognizer — Python dependencies
# Install with: pip install -r requirements.txt

pytesseract==0.3.13     # Python wrapper for Tesseract OCR engine
Pillow==10.4.0          # Image loading and pre-processing
requests==2.32.3        # Online RxNorm verification (optional but recommended)
